* origin/5.x:
  Jenkins bump to 5.0.0-preview-3-build.14
  cleanup of menubar
  using alloy api changes
  AP-323: add unbind return function
  fix tsc issues witth boulder merge, mostly Objects.readOptFrom
  AP-323: disable context menus, context toolbars, and toolbar items for editing images that are not editable
  AP-366: Updated the emoticons and charmap autocompleter to specify they need at minimum 2 chars before triggering
  TINY-1900: Updated the export to not use default
  TINY-1900: Replaced hasOwnContent with Obj.has
  TINY-1900: Renamed a couple function names to be more accurate
  TINY-1900: Added types, moved some functions into its own module and fixed other issues found during review
  TINY-1900: Fixed an issue where inline formatting would cause the delete key to do nothing
  AP-366: Added some missed types
  AP-366: Updated the emoticons & charmap autocompleter to only show if it's at the start of a word
  AP-366: Added "minChars" and "matches" properties to the autocompleter API that allows the autocompleter to determine if it should show based on the context when triggered
  TINY-1897: Fixed an issue that was causing incorrect viewport calculations for fixed position UI elements
  Fixing tests to work with proposed agar, and proposed alloy
  updating to match new alloy

# Conflicts:
#	src/themes/silver/main/ts/ui/general/Notification.ts

+++ src/core/main/ts/delete/InlineFormatDelete.ts
@@ -52,7 +52,7 @@ const deleteCaret = function (editor, forward) {
 
   return Arr.last(parentInlines).map(function (target) {
     const fromPos = CaretPosition.fromRangeStart(editor.selection.getRng());
-    if (DeleteUtils.willDeleteLastPositionInElement(forward, fromPos, target.dom())) {
+    if (DeleteUtils.willDeleteLastPositionInElement(forward, fromPos, target.dom()) && !CaretFormat.isEmptyCaretFormatElement(target)) {
       deleteLastPosition(forward, editor, target, parentInlines);
       return true;
     } else {
